fix: currently error with backspase and delet
