+++ src/main/script/interaction.js
@@ -1,12 +1,21 @@
+import {
+  setAtrsforKeys,
+  rmTxt,
+  toggleLanguage,
+  checkActive,
+  toggleActive,
+  getLowerCase,
+  getUpperCase,
+} from './helper';
+
 window.addEventListener('load', () => {
   const keyBoardWrap = document.querySelector('.keyboard_wrapp');
   const txtField = document.querySelector('.text-field');
   const keys = document.querySelectorAll('.keys');
   let focus = false;
-  let highlight = '';
   let language = 'en';
 
-  setAtrsforKeys(keys);
+  setAtrsforKeys();
 
   const space = document.querySelector('.space_key');
   space.setAttribute('keyname', 'Space');
@@ -56,7 +65,7 @@ window.addEventListener('load', () => {
   right.classList.add('arr');
 
   if (localStorage.getItem('active-language') && language !== localStorage.getItem('active-language')) {
-    for (let i = 0; i < keys.length; i++) {
+    for (let i = 0; i < keys.length; i += 1) {
       if (keys[i].getAttribute('keyRu')) {
         keys[i].innerHTML = keys[i].getAttribute('keyRu');
       }
@@ -72,7 +81,7 @@ window.addEventListener('load', () => {
       keyname = null;
     }
 
-    for (let i = 0; i < keys.length; i++) {
+    for (let i = 0; i < keys.length; i += 1) {
       if (
         e.key === keys[i].getAttribute('keyEn')
         || e.key === keys[i].getAttribute('keyRu')
@@ -98,7 +107,7 @@ window.addEventListener('load', () => {
   });
 
   window.addEventListener('keyup', (e) => {
-    for (let i = 0; i < keys.length; i++) {
+    for (let i = 0; i < keys.length; i += 1) {
       if (
         e.key === keys[i].getAttribute('keyEn')
         || e.key === keys[i].getAttribute('keyRu')
@@ -172,12 +181,14 @@ window.addEventListener('load', () => {
       if (e.target.classList.contains('left')) {
         if (txtField.selectionStart > 0) {
           const ind = txtField.selectionStart;
-          txtField.selectionStart = txtField.selectionEnd = ind - 1;
+          txtField.selectionStart = ind - 1;
+          txtField.selectionEnd = ind - 1;
         }
       } else if (e.target.classList.contains('right')) {
         const ind = txtField.selectionStart;
         if (ind < txtField.value.length) {
-          txtField.selectionStart = txtField.selectionEnd = ind + 1;
+          txtField.selectionStart = ind + 1;
+          txtField.selectionEnd = ind + 1;
         }
       }
     }
@@ -195,192 +206,29 @@ window.addEventListener('load', () => {
         const startStr = txtField.value.slice(0, [txtField.selectionEnd]);
         const endStr = txtField.value.slice([txtField.selectionEnd]);
         txtField.value = `${startStr} ${endStr}`;
-        txtField.selectionStart = txtField.selectionEnd = ind;
+        txtField.selectionStart = ind;
+        txtField.selectionEnd = ind;
       }
     } else if (target.length === 1) {
       txtField.value += target;
     } else if (keyname === 'Tab') {
       txtField.value += '    ';
     } else if (keyname === 'Backspace') {
-      if (txtField.value.length > 0) {
-        if (highlight === '') {
-          txtField.value = rmTxt(txtField.selectionEnd - 1, txtField.value);
-        } else {
-          txtField.value = rmTxt(highlight, txtField.value);
-          highlight = '';
-        }
+      if (txtField.value.length > 0 && txtField.selectionEnd > 0) {
+        const ind = txtField.selectionEnd - 1;
+        txtField.value = rmTxt(txtField, txtField.selectionEnd - 1, txtField.selectionEnd);
+        txtField.selectionStart = ind;
+        txtField.selectionEnd = ind;
       }
     } else if (keyname === 'Enter') {
       txtField.value += '\n';
     } else if (keyname === 'Delete') {
       if (txtField.value.length > 0 && txtField.selectionEnd < txtField.value.length) {
-        txtField.value = rmTxt(txtField.selectionEnd, txtField.value);
+        const ind = txtField.selectionEnd;
+        txtField.value = rmTxt(txtField, ind, ind + 1);
+        txtField.selectionStart = ind;
+        txtField.selectionEnd = ind;
       }
     }
   });
-
-  document.onselectionchange = () => {
-    //console.log(txtField.selectionStart);
-    //console.log(txtField.selectionEnd);
-  };
 });
-
-function rmTxt(target, txtValue) {
-  if (typeof target === 'number') {
-    return txtValue.slice(0, target);
-  }
-  const regExp = /target/i;
-  return txtValue.replace(regExp, '');
-}
-
-function getUpperCase(keys, key) {
-  if (key === 'ShiftLeft' || key === 'ShiftLeft') {
-    for (let i = 0; i < keys.length; i++) {
-      if (localStorage.getItem('active-language') === 'en') {
-        if (keys[i].getAttribute('keyEn') || keys[i].getAttribute('key')) {
-          keys[i].innerHTML = keys[i].getAttribute('UpperCaseNameEn');
-        }
-      } else if (keys[i].getAttribute('keyRu') || keys[i].getAttribute('key')) {
-        keys[i].innerHTML = keys[i].getAttribute('UpperCaseNameRu');
-      }
-    }
-  } else {
-    for (let i = 0; i < keys.length; i++) {
-      if (localStorage.getItem('active-language') === 'en') {
-        if (!keys[i].getAttribute('only-shift')) {
-          if (keys[i].getAttribute('keyEn') || keys[i].getAttribute('key')) {
-            keys[i].innerHTML = keys[i].getAttribute('UpperCaseNameEn');
-          }
-        }
-      } else if (!keys[i].getAttribute('only-shift') || keys[i].getAttribute('only-shift') === 'en') {
-        if (keys[i].getAttribute('keyRu') || keys[i].getAttribute('key')) {
-          keys[i].innerHTML = keys[i].getAttribute('UpperCaseNameRu');
-        }
-      }
-    }
-  }
-}
-
-function getLowerCase(keys) {
-  for (let i = 0; i < keys.length; i++) {
-    if (keys[i].getAttribute('key')) {
-      keys[i].innerHTML = keys[i].getAttribute('key');
-    }
-    if (localStorage.getItem('active-language') === 'en') {
-      if (keys[i].getAttribute('keyEn')) {
-        keys[i].innerHTML = keys[i].getAttribute('keyEn');
-      }
-    } else if (keys[i].getAttribute('keyRu')) {
-      keys[i].innerHTML = keys[i].getAttribute('keyRu');
-    }
-  }
-}
-
-function toggleActive(el) {
-  if (el.classList.contains('active')) {
-    el.classList.remove('active');
-  } else {
-    el.classList.add('active');
-  }
-}
-
-function checkActive(el) {
-  if (el.classList.contains('active')) {
-    return true;
-  }
-  return false;
-}
-
-function toggleLanguage(keys) {
-  if (localStorage.getItem('active-language') === 'en') {
-    for (let i = 0; i < keys.length; i++) {
-      if (keys[i].getAttribute('keyEn')) {
-        keys[i].innerHTML = keys[i].getAttribute('keyRu');
-      }
-    }
-    localStorage.setItem('active-language', 'ru');
-  } else {
-    for (let i = 0; i < keys.length; i++) {
-      if (keys[i].getAttribute('keyEn')) {
-        keys[i].innerHTML = keys[i].getAttribute('keyEn');
-      }
-    }
-    localStorage.setItem('active-language', 'en');
-  }
-}
-
-function setAtrsforKeys(keys) {
-  for (let i = 0; i < keys.length; i++) {
-    if (keys[i].getAttribute('keyEn')) {
-      keys[i].setAttribute(
-        'UpperCaseNameEn',
-        keys[i].getAttribute('keyEn').toUpperCase(),
-      );
-      keys[i].setAttribute(
-        'UpperCaseNameRu',
-        keys[i].getAttribute('keyRu').toUpperCase(),
-      );
-    }
-  }
-
-  keys[1].setAttribute('key', '1');
-  keys[2].setAttribute('key', '2');
-  keys[3].setAttribute('key', '3');
-  keys[4].setAttribute('key', '4');
-  keys[5].setAttribute('key', '5');
-  keys[6].setAttribute('key', '6');
-  keys[7].setAttribute('key', '7');
-  keys[8].setAttribute('key', '8');
-  keys[9].setAttribute('key', '9');
-  keys[10].setAttribute('key', '0');
-  keys[11].setAttribute('key', '-');
-  keys[12].setAttribute('key', '=');
-
-  keys[0].setAttribute('UpperCaseNameEn', '~');
-  keys[1].setAttribute('UpperCaseNameEn', '!');
-  keys[2].setAttribute('UpperCaseNameEn', '@');
-  keys[3].setAttribute('UpperCaseNameEn', '#');
-  keys[4].setAttribute('UpperCaseNameEn', '$');
-  keys[5].setAttribute('UpperCaseNameEn', '%');
-  keys[6].setAttribute('UpperCaseNameEn', '^');
-  keys[7].setAttribute('UpperCaseNameEn', '&');
-  keys[8].setAttribute('UpperCaseNameEn', '*');
-  keys[9].setAttribute('UpperCaseNameEn', '(');
-  keys[10].setAttribute('UpperCaseNameEn', ')');
-  keys[11].setAttribute('UpperCaseNameEn', '_');
-  keys[12].setAttribute('UpperCaseNameEn', '+');
-
-  keys[1].setAttribute('UpperCaseNameRu', '!');
-  keys[2].setAttribute('UpperCaseNameRu', '"');
-  keys[3].setAttribute('UpperCaseNameRu', '№');
-  keys[4].setAttribute('UpperCaseNameRu', ';');
-  keys[5].setAttribute('UpperCaseNameRu', '%');
-  keys[6].setAttribute('UpperCaseNameRu', ':');
-  keys[7].setAttribute('UpperCaseNameRu', '?');
-  keys[8].setAttribute('UpperCaseNameRu', '*');
-  keys[9].setAttribute('UpperCaseNameRu', '(');
-  keys[10].setAttribute('UpperCaseNameRu', ')');
-  keys[11].setAttribute('UpperCaseNameRu', '_');
-  keys[12].setAttribute('UpperCaseNameRu', '+');
-
-  keys[25].setAttribute('UpperCaseNameEn', '{');
-  keys[26].setAttribute('UpperCaseNameEn', '}');
-
-  keys[27].setAttribute('key', '\\');
-  keys[27].setAttribute('UpperCaseNameEn', '|');
-  keys[27].setAttribute('UpperCaseNameRu', '|');
-
-  keys[39].setAttribute('UpperCaseNameEn', ':');
-  keys[40].setAttribute('UpperCaseNameEn', '"');
-
-  keys[51].setAttribute('UpperCaseNameEn', '<');
-  keys[52].setAttribute('UpperCaseNameEn', '>');
-
-  keys[43].setAttribute('key', '/');
-  keys[43].setAttribute('UpperCaseNameEn', '/');
-  keys[43].setAttribute('UpperCaseNameRu', ',');
-
-  keys[53].setAttribute('key', '/');
-  keys[53].setAttribute('UpperCaseNameEn', '?');
-  keys[53].setAttribute('UpperCaseNameRu', '.');
-}
